Rename crank-editable to crankeditable
