+++ crankeditable/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@b9g/crank-editable",
+  "name": "@b9g/crankeditable",
   "version": "0.1.0",
   "license": "MIT",
   "type": "module",
